removed routing classes in favor of bitexpert/pathfinder refactored code to use bitexpert/pathfinder
