+++ composer.json
@@ -21,7 +21,8 @@
     "zendframework/zend-stratigility": "^1.1.0",
     "twig/twig": "v1.23.1",
     "willdurand/negotiation": "v2.0.1",
-    "bitexpert/slf4psrlog": "^0.1.0"
+    "bitexpert/slf4psrlog": "^0.1.0",
+    "bitexpert/pathfinder": "^0.1.0"
   },
   "require-dev": {
     "phpunit/php-code-coverage": "^2.1.0",
